% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-solutions.R
\docType{data}
\name{aoc17_solutions}
\alias{aoc17_solutions}
\title{Advent of Code 2017 solutions}
\format{An object of class \code{list} of length 49.}
\usage{
aoc17_solutions
}
\description{
Advent of Code 2017 solutions
}
\keyword{datasets}
