% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{checksum_range}
\alias{checksum_range}
\alias{checksum_evenly_divisible}
\title{Day 2: Corruption Checksum}
\usage{
checksum_range(x)

checksum_evenly_divisible(x)
}
\description{
Corruption Checksum
}
\details{
To quote the \href{http://adventofcode.com/2017/day/2}{website}

\strong{Part One}

As you walk through the door, a glowing humanoid shape yells in your
direction. "You there! Your state appears to be idle. Come help us repair the
corruption in this spreadsheet - if we take another millisecond, we'll have
to display an hourglass cursor!"

The spreadsheet consists of rows of apparently-random numbers. To make sure
the recovery process is on the right track, they need you to calculate the
spreadsheet's checksum. For each row, determine the difference between the
largest value and the smallest value; the checksum is the sum of all of these
differences.

For example, given the following spreadsheet:\preformatted{5 1 9 5
7 5 3
2 4 6 8
}
\itemize{
\item The first row's largest and smallest values are 9 and 1, and their
difference is 8.
\item The second row's largest and smallest values are 7 and 3, and their
difference is 4.
\item The third row's difference is 6.
}

In this example, the spreadsheet's checksum would be 8 + 4 + 6 = 18.

\strong{Part Two}

"Based on what we're seeing, it looks like all the User wanted is some
information about the evenly divisible values in the spreadsheet.
Unfortunately, none of us are equipped for that kind of calculation - most of
us specialize in bitwise operations."

It sounds like the goal is to find the only two numbers in each row where one
evenly divides the other - that is, where the result of the division
operation is a whole number. They would like you to find those numbers on
each line, divide them, and add up each line's result.

For example, given the following spreadsheet:\preformatted{5 9 2 8
9 4 7 3
3 8 6 5
}
\itemize{
\item In the first row, the only two numbers that evenly divide are 8 and 2;
the result of this division is 4.
\item In the second row, the two numbers are 9 and 3; the result is 3.
\item In the third row, the result is 2.
}

In this example, the sum of the results would be 4 + 3 + 2 = 9.
}
