% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{count_valid_passphrases}
\alias{count_valid_passphrases}
\alias{no_repeated_words}
\alias{no_anagrams}
\title{Day 4: High-Entropy Passphrases}
\usage{
count_valid_passphrases(x, rule = no_repeated_words)

no_repeated_words(x)

no_anagrams(x)
}
\description{
\href{http://adventofcode.com/2017/day/4}{High-Entropy Passphrases}
}
\details{
\strong{Part One}

A new system policy has been put in place that requires all accounts to
use a \emph{passphrase} instead of simply a pass\emph{word}. A passphrase consists
of a series of words (lowercase letters) separated by spaces.

To ensure security, a valid passphrase must contain no duplicate words.

For example:
\itemize{
\item \code{aa bb cc dd ee} is valid.
\item \code{aa bb cc dd aa} is not valid - the word \code{aa} appears more than
once.
\item \code{aa bb cc dd aaa} is valid - \code{aa} and \code{aaa} count as different
words.
}

The system's full passphrase list is available as your puzzle input.
\emph{How many passphrases are valid?}

\strong{Part Two}

For added security, yet another system policy
has been put in place. Now, a valid passphrase must contain no two words
that are anagrams of each other - that is, a passphrase is invalid if
any word's letters can be rearranged to form any other word in the
passphrase.

For example:
\itemize{
\item \code{abcde fghij} is a valid passphrase.
\item \code{abcde xyz ecdab} is not valid - the letters from the third word can
be rearranged to form the first word.
\item \code{a ab abc abd abf abj} is a valid passphrase, because \emph{all} letters
need to be used when forming another word.
\item \code{iiii oiii ooii oooi oooo} is valid.
\item \code{oiii ioii iioi iiio} is not valid - any of these words can be
rearranged to form any other word.
}

Under this new system policy, \emph{how many passphrases are valid?}
}
