% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{follow_day05a_instructions}
\alias{follow_day05a_instructions}
\alias{follow_day05b_instructions}
\title{Day 5: A Maze of Twisty Trampolines, All Alike}
\usage{
follow_day05a_instructions(x)

follow_day05b_instructions(x)
}
\description{
\href{http://adventofcode.com/2017/day/5}{A Maze of Twisty Trampolines, All Alike}
}
\details{
\strong{Part One}

An urgent interrupt arrives from the CPU: it's trapped in a maze of jump
instructions, and it would like assistance from any programs with spare
cycles to help find the exit.

The message includes a list of the offsets for each jump. Jumps are
relative: \code{-1} moves to the previous instruction, and \code{2} skips the next
one. Start at the first instruction in the list. The goal is to follow
the jumps until one leads \emph{outside} the list.

In addition, these instructions are a little strange; after each jump,
the offset of that instruction increases by \code{1}. So, if you come across
an offset of \code{3}, you would move three instructions forward, but change
it to a \code{4} for the next time it is encountered.

For example, consider the following list of jump offsets:\preformatted{0
3
0
1
-3
}

Positive jumps ("forward") move downward; negative jumps move upward.
For legibility in this example, these offset values will be written all
on one line, with the current instruction marked in parentheses. The
following steps would be taken before an exit is found:
\itemize{
\item \code{(0) 3  0  1  -3 } - \emph{before} we have taken any steps.
\item \code{(1) 3  0  1  -3 } - jump with offset \code{0} (that is, don't jump at
all). Fortunately, the instruction is then incremented to \code{1}.
\item \code{ 2 (3) 0  1  -3 } - step forward because of the instruction we just
modified. The first instruction is incremented again, now to \code{2}.
\item \code{ 2  4  0  1 (-3)} - jump all the way to the end; leave a \code{4}
behind.
\item \code{ 2 (4) 0  1  -2 } - go back to where we just were; increment \code{-3}
to \code{-2}.
\item \code{ 2  5  0  1  -2 } - jump \code{4} steps forward, escaping the maze.
}

In this example, the exit is reached in \code{5} steps.

\emph{How many steps} does it take to reach the exit?

\strong{Part Two}

Now, the jumps are even stranger: after each jump, if the offset was
\emph{three or more}, instead \emph{decrease} it by \code{1}. Otherwise, increase it by
\code{1} as before.

Using this rule with the above example, the process now takes \code{10}
steps, and the offset values after finding the exit are left as
\code{2 3 2 3 -1}.

\emph{How many steps} does it now take to reach the exit?
}
