% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{analyze_reallocations}
\alias{analyze_reallocations}
\title{Day 6: Title}
\usage{
analyze_reallocations(xs)
}
\description{
\href{http://adventofcode.com/2017/day/6}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
