% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{twist_knot}
\alias{twist_knot}
\alias{knot_hash}
\title{Day 10: Knot Hash}
\usage{
twist_knot(args)

knot_hash(seed)
}
\arguments{
\item{args}{a list of arguments to control the twisting}

\item{seed}{a value to hash}
}
\description{
\href{http://adventofcode.com/2017/day/10}{Knot Hash}
}
\details{
\strong{Part One}

You come across some programs that are trying to implement a software
emulation of a hash based on knot-tying. The hash these programs are
implementing isn't very strong, but you decide to help them anyway. You
make a mental note to remind the Elves later not to invent their own
cryptographic functions.

This hash function simulates tying a knot in a circle of string with 256
marks on it. Based on the input to be hashed, the function repeatedly
selects a span of string, brings the ends together, and gives the span a
half-twist to reverse the order of the marks within it. After doing this
many times, the order of the marks is used to build the resulting hash.\preformatted{  4--5   pinch   4  5           4   1
 /    \  5,0,1  / \/ \  twist  / \ / \
3      0  -->  3      0  -->  3   X   0
 \    /         \ /\ /         \ / \ /
  2--1           2  1           2   5
}

To achieve this, begin with a \emph{list} of numbers from \code{0} to \code{255}, a
\emph{current position} which begins at \code{0} (the first element in the list),
a \emph{skip size} (which starts at \code{0}), and a sequence of \emph{lengths} (your
puzzle input). Then, for each length:
\itemize{
\item \emph{Reverse} the order of that \emph{length} of elements in the \emph{list},
starting with the element at the \emph{current position}.
\item \emph{Move} the \emph{current position} forward by that \emph{length} plus the
\emph{skip size}.
\item \emph{Increase} the \emph{skip size} by one.
}

The \emph{list} is circular; if the \emph{current position} and the \emph{length} try
to reverse elements beyond the end of the list, the operation reverses
using as many extra elements as it needs from the front of the list. If
the \emph{current position} moves past the end of the list, it wraps around
to the front. \emph{Lengths} larger than the size of the \emph{list} are invalid.

Here's an example using a smaller list:

Suppose we instead only had a circular list containing five elements,
\code{0, 1, 2, 3, 4}, and were given input lengths of \code{3, 4, 1, 5}.
\itemize{
\item The list begins as \code{[0] 1 2 3 4} (where square brackets indicate the
\emph{current position}).
\item The first length, \code{3}, selects \code{([0] 1 2) 3 4} (where parentheses
indicate the sublist to be reversed).
\item After reversing that section (\code{0 1 2} into \code{2 1 0}), we get
\code{([2] 1 0) 3 4}.
\item Then, the \emph{current position} moves forward by the \emph{length}, \code{3},
plus the \emph{skip size}, 0: \code{2 1 0 [3] 4}. Finally, the \emph{skip size}
increases to \code{1}.
\item The second length, \code{4}, selects a section which wraps:
\code{2 1) 0 ([3] 4}.
\item The sublist \code{3 4 2 1} is reversed to form \code{1 2 4 3}:
\code{4 3) 0 ([1] 2}.
\item The \emph{current position} moves forward by the \emph{length} plus the \emph{skip
size}, a total of \code{5}, causing it not to move because it wraps
around: \code{4 3 0 [1] 2}. The \emph{skip size} increases to \code{2}.
\item The third length, \code{1}, selects a sublist of a single element, and so
reversing it has no effect.
\item The \emph{current position} moves forward by the \emph{length} (\code{1}) plus the
\emph{skip size} (\code{2}): \code{4 [3] 0 1 2}. The \emph{skip size} increases to \code{3}.
\item The fourth length, \code{5}, selects every element starting with the
second: \code{4) ([3] 0 1 2}. Reversing this sublist (\code{3 0 1 2 4} into
\code{4 2 1 0 3}) produces: \code{3) ([4] 2 1 0}.
\item Finally, the \emph{current position} moves forward by \code{8}: \code{3 4 2 1 [0]}.
The \emph{skip size} increases to \code{4}.
}

In this example, the first two numbers in the list end up being \code{3} and
\code{4}; to check the process, you can multiply them together to produce
\code{12}.

However, you should instead use the standard list size of \code{256} (with
values \code{0} to \code{255}) and the sequence of \emph{lengths} in your puzzle input.
Once this process is complete, \emph{what is the result of multiplying the
first two numbers in the list}?

\strong{Part Two}

The logic you've constructed forms a single \emph{round} of the \emph{Knot Hash}
algorithm; running the full thing requires many of these rounds. Some
input and output processing is also required.

First, from now on, your input should be taken not as a list of numbers,
but as a string of bytes instead. Unless otherwise specified, convert
characters to bytes using their \href{https://en.wikipedia.org/wiki/ASCII#Printable_characters}{ASCII codes}. This
will allow you to handle arbitrary ASCII strings, and it also ensures
that your input lengths are never larger than \code{255}. For example, if you
are given \code{1,2,3}, you should convert it to the ASCII codes for each
character: \code{49,44,50,44,51}.

Once you have determined the sequence of lengths to use, add the
following lengths to the end of the sequence: \code{17, 31, 73, 47, 23}. For
example, if you are given \code{1,2,3}, your final sequence of lengths should
be \code{49,44,50,44,51,17,31,73,47,23} (the ASCII codes from the input
string combined with the standard length suffix values).

Second, instead of merely running one \emph{round} like you did above, run a
total of \code{64} rounds, using the same \emph{length} sequence in each round.
The \emph{current position} and \emph{skip size} should be preserved between
rounds. For example, if the previous example was your first round, you
would start your second round with the same \emph{length} sequence
(\code{3, 4, 1, 5, 17, 31, 73, 47, 23}, now assuming they came from ASCII
codes and include the suffix), but start with the previous round's
\emph{current position} (\code{4}) and \emph{skip size} (\code{4}).

Once the rounds are complete, you will be left with the numbers from \code{0}
to \code{255} in some order, called the \emph{sparse hash}. Your next task is to
reduce these to a list of only \code{16} numbers called the \emph{dense hash}. To
do this, use numeric bitwise
\href{https://en.wikipedia.org/wiki/Bitwise_operation#XOR}{XOR} to combine
each consecutive block of \code{16} numbers in the sparse hash (there are
\code{16} such blocks in a list of \code{256} numbers). So, the first element in
the dense hash is the first sixteen elements of the sparse hash XOR'd
together, the second element in the dense hash is the second sixteen
elements of the sparse hash XOR'd together, etc.

For example, if the first sixteen elements of your sparse hash are as
shown below, and the XOR operator is \code{^}, you would calculate the first
output number like this:\preformatted{65 ^ 27 ^ 9 ^ 1 ^ 4 ^ 3 ^ 40 ^ 50 ^ 91 ^ 7 ^ 6 ^ 0 ^ 2 ^ 5 ^ 68 ^ 22 = 64
}

Perform this operation on each of the sixteen blocks of sixteen numbers
in your sparse hash to determine the sixteen numbers in your dense hash.

Finally, the standard way to represent a Knot Hash is as a single
\href{https://en.wikipedia.org/wiki/Hexadecimal}{hexadecimal} string; the
final output is the dense hash in hexadecimal notation. Because each
number in your dense hash will be between \code{0} and \code{255} (inclusive),
always represent each number as two hexadecimal digits (including a
leading zero as necessary). So, if your first three numbers are
\code{64, 7, 255}, they correspond to the hexadecimal numbers \code{40, 07, ff},
and so the first six characters of the hash would be \code{4007ff}. Because
every Knot Hash is sixteen such numbers, the hexadecimal representation
is always \code{32} hexadecimal digits (\code{0}-\code{f}) long.

Here are some example hashes:
\itemize{
\item The empty string becomes \code{a2582a3a0e66e6e86e3812dcb672a272}.
\item \code{AoC 2017} becomes \code{33efeb34ea91902bb2f59c9920caa6cd}.
\item \code{1,2,3} becomes \code{3efbe78a8d82f29979031a4aa0b16a9d}.
\item \code{1,2,4} becomes \code{63960835bcdc130f0b66d7ff4f6a5a8e}.
}

Treating your puzzle input as a string of ASCII characters, \emph{what is the
Knot Hash of your puzzle input?} Ignore any leading or trailing
whitespace you might encounter.
}
\examples{
args <- list(
  items = 0:4,
  step = 0,
  current_position = 1,
  lengths = c(3, 4, 1, 5))
twist_knot(args)
knot_hash("test")
}
