% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day11.R
\name{hexagon_distance}
\alias{hexagon_distance}
\alias{find_longest_hexagon_distance}
\title{Day 11: Hex Ed}
\usage{
hexagon_distance(steps)

find_longest_hexagon_distance(steps)
}
\description{
Hex Ed
}
\details{
To quote the \href{http://adventofcode.com/2017/day/11}{website}

\strong{Part One}

Crossing the bridge, you've barely reached the other side of the stream
when a program comes up to you, clearly in distress. "It's my child
process," she says, "he's gotten lost in an infinite grid!"

Fortunately for her, you have plenty of experience with infinite grids.

Unfortunately for you, it's a \href{https://en.wikipedia.org/wiki/Hexagonal_tiling}{hex grid}.

The hexagons ("hexes") in this grid are aligned such that adjacent hexes can
be found to the north, northeast, southeast, south, southwest, and northwest:\preformatted{  \ n  /
nw +--+ ne
  /    \
-+      +-
  \    /
sw +--+ se
  / s  \
}

You have the path the child process took. Starting where he started, you
need to determine the fewest number of steps required to reach him. (A
"step" means to move from the hex you are in to any adjacent hex.)

For example:
\itemize{
\item \code{ne,ne,ne} is \code{3} steps away.
\item \code{ne,ne,sw,sw} is \code{0} steps away (back where you started).
\item \code{ne,ne,s,s} is \code{2} steps away (\code{se,se}).
\item \code{se,sw,se,sw,sw} is \code{3} steps away (\code{s,s,sw}).
}

\strong{Part Two}

\emph{How many steps away} is the \emph{furthest} he ever got from his starting
position?
}
