% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{generate_grid_hashes}
\alias{generate_grid_hashes}
\alias{str_sum_ones}
\alias{count_grid_regions}
\title{Day 14: Disk Defragmentation}
\usage{
generate_grid_hashes(key, seq = 0:127)

str_sum_ones(string)

count_grid_regions(bits)
}
\arguments{
\item{key}{a seed for the hashing}

\item{seq}{a set of suffix numbers for the keys}

\item{string}{a string}

\item{bits}{a matrix of bits to flood-fill}
}
\description{
\href{http://adventofcode.com/2017/day/14}{Disk Defragmentation}
}
\details{
\strong{Part One}

Suddenly, a scheduled job activates the system's \href{https://en.wikipedia.org/wiki/Defragmentation}{disk defragmenter}. Were the
situation different, you might \href{https://www.youtube.com/watch?v=kPv1gQ5Rs8A&t=37}{sit and watch it for a while}, but today, you
just don't have that kind of time. It's soaking up valuable system
resources that are needed elsewhere, and so the only option is to help
it finish its task as soon as possible.

The disk in question consists of a 128x128 grid; each square of the grid
is either \emph{free} or \emph{used}. On this disk, the state of the grid is
tracked by the bits in a sequence of \href{http://adventofcode.com/2017/day/10}{knot hashes}.

A total of 128 knot hashes are calculated, each corresponding to a
single row in the grid; each hash contains 128 bits which correspond to
individual grid squares. Each bit of a hash indicates whether that
square is \emph{free} (\code{0}) or \emph{used} (\code{1}).

The hash inputs are a key string (your puzzle input), a dash, and a
number from \code{0} to \code{127} corresponding to the row. For example, if your
key string were \code{flqrgnkx}, then the first row would be given by the
bits of the knot hash of \code{flqrgnkx-0}, the second row from the bits of
the knot hash of \code{flqrgnkx-1}, and so on until the last row,
\code{flqrgnkx-127}.

The output of a knot hash is traditionally represented by 32 hexadecimal
digits; each of these digits correspond to 4 bits, for a total of
\code{4 * 32 = 128} bits. To convert to bits, turn each hexadecimal digit to
its equivalent binary value, high-bit first: \code{0} becomes \code{0000}, \code{1}
becomes \code{0001}, \code{e} becomes \code{1110}, \code{f} becomes \code{1111}, and so on; a
hash that begins with \code{a0c2017...} in hexadecimal would begin with
\code{10100000110000100000000101110000...} in binary.

Continuing this process, the \emph{first 8 rows and columns} for key
\code{flqrgnkx} appear as follows, using \code{#} to denote used squares, and \code{.}
to denote free ones:\preformatted{##.#.#..-->
.#.#.#.#
....#.#.
#.#.##.#
.##.#...
##..#..#
.#...#..
##.#.##.-->
|      |
V      V
}

In this example, \code{8108} squares are used across the entire 128x128 grid.

Given your actual key string, \emph{how many squares are used}?

\strong{Part Two}

Now, all the defragmenter needs to know is the number
of \emph{regions}. A region is a group of \emph{used} squares that are all
\emph{adjacent}, not including diagonals. Every used square is in exactly one
region: lone used squares form their own isolated regions, while several
adjacent squares all count as a single region.

In the example above, the following nine regions are visible, each
marked with a distinct digit:\preformatted{11.2.3..-->
.1.2.3.4
....5.6.
7.8.55.9
.88.5...
88..5..8
.8...8..
88.8.88.-->
|      |
V      V
}

Of particular interest is the region marked \code{8}; while it does not
appear contiguous in this small view, all of the squares marked \code{8} are
connected when considering the whole 128x128 grid. In total, in this
example, \code{1242} regions are present.

\emph{How many regions} are present given your key string?
}
\examples{
test_grid <- "flqrgnkx" \%>\%
  generate_grid_hashes(0:7)
str_sum_ones(test_grid)
count_grid_regions(test_grid)
}
