% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day20.R
\name{find_slowest_particle}
\alias{find_slowest_particle}
\alias{create_particles}
\alias{search_and_destroy_particles}
\title{Day 20: Title}
\usage{
find_slowest_particle(p_strings)

create_particles(p_strings)

search_and_destroy_particles(particles)
}
\description{
\href{http://adventofcode.com/2017/day/20}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
