% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day22.R
\name{day22}
\alias{day22}
\alias{virus_grid}
\alias{evolved_virus_grid}
\title{Day 22: Sporifica Virus}
\usage{
virus_grid(start)

evolved_virus_grid(start)
}
\arguments{
\item{start}{starting start of grid (one string per row)}
}
\description{
\href{http://adventofcode.com/2017/day/22}{Sporifica Virus}
}
\details{
\strong{Part One}

Diagnostics indicate that the local \emph{grid computing cluster} has been
contaminated with the \emph{Sporifica Virus}. The grid computing cluster is a
seemingly-infinite
two-dimensional grid of compute nodes. Each node is either \emph{clean} or
\emph{infected} by the virus.

To \href{https://en.wikipedia.org/wiki/Morris_worm#The_mistake}{prevent overloading} the
nodes (which would render them useless to the virus) or detection by
system administrators, exactly one \emph{virus carrier} moves through the
network, infecting or cleaning nodes as it moves. The virus carrier is
always located on a single node in the network (the \emph{current node}) and
keeps track of the \emph{direction} it is facing.

To avoid detection, the virus carrier works in bursts; in each burst, it
\emph{wakes up}, does some \emph{work}, and goes back to \emph{sleep}. The following
steps are all executed \emph{in order} one time each burst:
\itemize{
\item If the \emph{current node} is \emph{infected}, it turns to its \emph{right}.
Otherwise, it turns to its \emph{left}. (Turning is done in-place; the
\emph{current node} does not change.)
\item If the \emph{current node} is \emph{clean}, it becomes \emph{infected}. Otherwise,
it becomes \emph{cleaned}. (This is done \emph{after} the node is considered
for the purposes of changing direction.)
\item The virus carrier
\href{https://www.youtube.com/watch?v=2vj37yeQQHg}{moves} \emph{forward} one
node in the direction it is facing.
}

Diagnostics have also provided a \emph{map of the node infection status}
(your puzzle input). \emph{Clean} nodes are shown as \code{.}; \emph{infected} nodes
are shown as \code{#}. This map only shows the center of the grid; there are
many more nodes beyond those shown, but none of them are currently
infected.

The virus carrier begins in the middle of the map facing \emph{up}.

For example, suppose you are given a map like this:\preformatted{..#
#..
...
}

Then, the middle of the infinite grid looks like this, with the virus
carrier's position marked with \code{[ ]}:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . # . . .
. . . #[.]. . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

The virus carrier is on a \emph{clean} node, so it turns \emph{left}, \emph{infects}
the node, and moves left:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . # . . .
. . .[#]# . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

The virus carrier is on an \emph{infected} node, so it turns \emph{right},
\emph{cleans} the node, and moves up:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . .[.]. # . . .
. . . . # . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

Four times in a row, the virus carrier finds a \emph{clean}, \emph{infects} it,
turns \emph{left}, and moves forward, ending in the same place and still
facing up:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . #[#]. # . . .
. . # # # . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

Now on the same node as before, it sees an infection, which causes it to
turn \emph{right}, \emph{clean} the node, and move forward:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . # .[.]# . . .
. . # # # . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

After the above actions, a total of \code{7} bursts of activity had taken
place. Of them, \code{5} bursts of activity caused an infection.

After a total of \code{70}, the grid looks like this, with the virus carrier
facing up:\preformatted{. . . . . # # . .
. . . . # . . # .
. . . # . . . . #
. . # . #[.]. . #
. . # . # . . # .
. . . . . # # . .
. . . . . . . . .
. . . . . . . . .
}

By this time, \code{41} bursts of activity caused an infection (though most
of those nodes have since been cleaned).

After a total of \code{10000} bursts of activity, \code{5587} bursts will have
caused an infection.

Given your actual map, after \code{10000} bursts of activity, \emph{how many
bursts cause a node to become infected}? (Do not count nodes that begin
infected.)

\strong{Part Two}

As you go to remove the virus from the infected nodes, it \emph{evolves} to
resist your attempt.

Now, before it infects a clean node, it will \emph{weaken} it to disable your
defenses. If it encounters an infected node, it will instead \emph{flag} the
node to be cleaned in the future. So:
\itemize{
\item \emph{Clean} nodes become \emph{weakened}.
\item \emph{Weakened} nodes become \emph{infected}.
\item \emph{Infected} nodes become \emph{flagged}.
\item \emph{Flagged} nodes become \emph{clean}.
}

Every node is always in exactly one of the above states.

The virus carrier still functions in a similar way, but now uses the
following logic during its bursts of action:
\itemize{
\item Decide which way to turn based on the \emph{current node}:
\itemize{
\item If it is \emph{clean}, it turns \emph{left}.
\item If it is \emph{weakened}, it does \emph{not} turn, and will continue
moving in the same direction.
\item If it is \emph{infected}, it turns \emph{right}.
\item If it is \emph{flagged}, it \emph{reverses} direction, and will go back
the way it came.
}
\item Modify the state of the \emph{current node}, as described above.
\item The virus carrier moves \emph{forward} one node in the direction it is
facing.
}

Start with the same map (still using \code{.} for \emph{clean} and \code{#} for
infected) and still with the virus carrier starting in the middle and
facing \emph{up}.

Using the same initial state as the previous example, and drawing
\emph{weakened} as \code{W} and \emph{flagged} as \code{F}, the middle of the infinite grid
looks like this, with the virus carrier's position again marked with
\code{[ ]}:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . # . . .
. . . #[.]. . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

This is the same as before, since no initial nodes are \emph{weakened} or
\emph{flagged}. The virus carrier is on a clean node, so it still turns left,
instead \emph{weakens} the node, and moves left:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . # . . .
. . .[#]W . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

The virus carrier is on an infected node, so it still turns right,
instead \emph{flags} the node, and moves up:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . .[.]. # . . .
. . . F W . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

This process repeats three more times, ending on the previously-flagged
node and facing right:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . W W . # . . .
. . W[F]W . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

Finding a flagged node, it reverses direction and \emph{cleans} the node:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . W W . # . . .
. .[W]. W . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

The \emph{weakened} node becomes infected, and it continues in the same
direction:\preformatted{. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
. . W W . # . . .
.[.]# . W . . . .
. . . . . . . . .
. . . . . . . . .
. . . . . . . . .
}

Of the first \code{100} bursts, \code{26} will result in \emph{infection}.
Unfortunately, another feature of this evolved virus is \emph{speed}; of the
first \code{10000000} bursts, \code{2511944} will result in \emph{infection}.

Given your actual map, after \code{10000000} bursts of activity, \emph{how many
bursts cause a node to become infected}? (Do not count nodes that begin
infected.)

[ ]: R:%20
[.]: R:.
[#]: R:#
[.]: R:.
[#]: R:#
[.]: R:.
[.]: R:.
[ ]: R:%20
[.]: R:.
[#]: R:#
[.]: R:.
[F]: R:F
[W]: R:W
[.]: R:.
}
\examples{
start <- c("..#", "#..", "...")
grid1 <- virus_grid(start)
grid1$step()

grid2 <- evolved_virus_grid(start)
grid2$step()
}
