% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{find_strongest_bridge}
\alias{find_strongest_bridge}
\alias{find_longest_bridge}
\alias{compute_bridge_strength}
\alias{compute_bridge_length}
\title{Day 24: Title}
\usage{
find_strongest_bridge(bridges)

find_longest_bridge(bridges)

compute_bridge_strength(bridge)

compute_bridge_length(bridge)
}
\description{
\href{http://adventofcode.com/2017/day/24}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
