% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day25.R
\name{turing_machine}
\alias{turing_machine}
\alias{create_tm_rule}
\title{Day 25: Title}
\usage{
turing_machine(rules, starting_state)

create_tm_rule(start, scanned, print, move, final)
}
\description{
\href{http://adventofcode.com/2017/day/25}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
