% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07-todo.R, R/dayxx-template.R
\name{find_root_program}
\alias{find_root_program}
\alias{dayxx_b_stub}
\alias{dayxx_a_stub}
\title{Day X: Title}
\usage{
find_root_program(lines)

dayxx_b_stub(x)

dayxx_a_stub(x)

dayxx_b_stub(x)
}
\description{
\href{http://adventofcode.com/2017/day/X}{Title}

\href{http://adventofcode.com/2017/day/X}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}

\strong{Part One}

\strong{Part Two}
}
