% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{int_to_n_bits}
\alias{int_to_n_bits}
\title{Convert a vector of integers to a vector of strings of n binary digits}
\usage{
int_to_n_bits(xs, n = 16, start = 1)
}
\arguments{
\item{xs}{a vector of integers}

\item{n}{number of the highest bit. Defaults to 16.}

\item{start}{number of the lowest bit. Defaults to 1, so all bits up to the
nth are selected.}
}
\value{
a character vector of binary strings
}
\description{
Convert a vector of integers to a vector of strings of n binary digits
}
