% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_if}
\alias{keep_if}
\title{Filter values}
\usage{
keep_if(data, predicate)
}
\arguments{
\item{data}{a list to filter}

\item{predicate}{a function for filtering items. Values that return \code{TRUE}
are kept.}
}
\value{
the items where the predicate function is \code{TRUE}
}
\description{
This function is \code{\link[base:Filter]{base::Filter()}} with its arguments reversed.
}
