% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{A negated version of \code{\\\%in\\\%}}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{values to be non-matched}

\item{table}{values to be non-matched against}
}
\description{
This function is a negated version of \code{\link[base]{\%in\%}}. Returns
\code{TRUE} when an element on the left-hand side does not have a match on the
right-hand side.
}
\examples{
c("a", "b", "x", "y") \%nin\% c("a", "b", "c", "d")
}
\keyword{internal}
