% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_text_lines}
\alias{read_text_lines}
\title{Split and trim new-line-delimited string of values}
\usage{
read_text_lines(x)
}
\arguments{
\item{x}{a string with values separated by newlines}
}
\value{
a character vector of values split at newlines and with leading and
trailing spaces removed
}
\description{
Split and trim new-line-delimited string of values
}
\examples{
read_text_lines("a\\n  b  \\nc    ")
}
