% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day18.R
\name{create_solo}
\alias{create_solo}
\alias{create_duet}
\title{Day 18: Title}
\usage{
create_solo(commands)

create_duet(program_id, commands)
}
\description{
\href{http://adventofcode.com/2017/day/18}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
