% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day19.R
\name{collect_board_letters}
\alias{collect_board_letters}
\alias{walk_board_from_start}
\alias{create_board}
\title{Day 19: Title}
\usage{
collect_board_letters(board)

walk_board_from_start(board)

create_board(strings)
}
\description{
\href{http://adventofcode.com/2017/day/19}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
