% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{day24}
\alias{day24}
\alias{find_strongest_bridge}
\alias{find_longest_bridge}
\alias{compute_bridge_strength}
\alias{compute_bridge_length}
\alias{connect_free_pieces}
\title{Day 24: Electromagnetic Moat}
\usage{
find_strongest_bridge(bridges)

find_longest_bridge(bridges)

compute_bridge_strength(bridge)

compute_bridge_length(bridge)

connect_free_pieces(piece, active_value, pieces)
}
\arguments{
\item{bridges}{a string describing a bridge}

\item{bridge}{a string describing a bridge}

\item{piece}{starting piece}

\item{active_value}{the value that needs to be matched}

\item{pieces}{vector of pieces to search for a match}
}
\description{
\href{http://adventofcode.com/2017/day/24}{Electromagnetic Moat}
}
\details{
\strong{Part One}

The CPU itself is a large, black building surrounded by a bottomless
pit. Enormous metal tubes extend outward from the side of the building
at regular intervals and descend down into the void. There's no way to
cross, but you need to get inside.

No way, of course, other than building a \emph{bridge} out of the magnetic
components strewn about nearby.

Each component has two \emph{ports}, one on each end. The ports come in all
different types, and only matching types can be connected. You take an
inventory of the components by their port types (your puzzle input).
Each port is identified by the number of \emph{pins} it uses; more pins mean
a stronger connection for your bridge. A \code{3/7} component, for example,
has a type-\code{3} port on one side, and a type-\code{7} port on the other.

Your side of the pit is metallic; a perfect surface to connect a
magnetic, \emph{zero-pin port}. Because of this, the first port you use must
be of type \code{0}. It doesn't matter what type of port you end with; your
goal is just to make the bridge as strong as possible.

The \emph{strength} of a bridge is the sum of the port types in each
component. For example, if your bridge is made of components \code{0/3},
\code{3/7}, and \code{7/4}, your bridge has a strength of \code{0+3 + 3+7 + 7+4 = 24}.

For example, suppose you had the following components:\preformatted{0/2
2/2
2/3
3/4
3/5
0/1
10/1
9/10
}

With them, you could make the following valid bridges:
\itemize{
\item \code{0/1}
\item \code{0/1}--\code{10/1}
\item \code{0/1}--\code{10/1}--\code{9/10}
\item \code{0/2}
\item \code{0/2}--\code{2/3}
\item \code{0/2}--\code{2/3}--\code{3/4}
\item \code{0/2}--\code{2/3}--\code{3/5}
\item \code{0/2}--\code{2/2}
\item \code{0/2}--\code{2/2}--\code{2/3}
\item \code{0/2}--\code{2/2}--\code{2/3}--\code{3/4}
\item \code{0/2}--\code{2/2}--\code{2/3}--\code{3/5}
}

(Note how, as shown by \code{10/1}, order of ports within a component doesn't
matter. However, you may only use each port on a component once.)

Of these bridges, the \emph{strongest} one is \code{0/1}--\code{10/1}--\code{9/10}; it has a
strength of \code{0+1 + 1+10 + 10+9 = 31}.

\emph{What is the strength of the strongest bridge you can make} with the
components you have available?

\strong{Part Two}

The bridge you've built isn't long enough; you can't jump the rest of
the way.

In the example above, there are two longest bridges:
\itemize{
\item \code{0/2}--\code{2/2}--\code{2/3}--\code{3/4}
\item \code{0/2}--\code{2/2}--\code{2/3}--\code{3/5}
}

Of them, the one which uses the \code{3/5} component is stronger; its
strength is \code{0+2 + 2+2 + 2+3 + 3+5 = 19}.

\emph{What is the strength of the longest bridge you can make?} If you can
make multiple bridges of the longest length, pick the \emph{strongest} one.
}
\examples{
pieces <- c("0/2", "2/2", "2/3",  "3/4", "3/5",  "0/1", "10/1", "9/10")

# Starts with empty piece ("") and request a piece with 0. Returns all
# possible full-length bridges from those starting conditions.
bridges <- connect_free_pieces("", 0, pieces)

find_strongest_bridge(bridges)

bridges \%>\%
  find_strongest_bridge() \%>\%
  compute_bridge_strength()

find_longest_bridge(bridges)

bridges \%>\%
  find_longest_bridge() \%>\%
  find_strongest_bridge()
}
