% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dayxx-template.R
\name{dayxx_a_stub}
\alias{dayxx_a_stub}
\alias{dayxx_b_stub}
\title{Day X: Title}
\usage{
dayxx_a_stub(x)

dayxx_b_stub(x)
}
\description{
\href{http://adventofcode.com/2017/day/X}{Title}
}
\details{
\strong{Part One}

\strong{Part Two}
}
